+++ src/drawings/impaledDrawing.ts
@@ -1,7 +1,9 @@
+import { CustomError } from "../interfaces/ErrorTypes";
 import { terminalColours } from "../utils/terminalColours";
-export function drawImpaled(error: any) {
-  let errorCode = error.code || error.status ? error.code || error.status : "xxx";
-  let errorString = error.message || error.text ? error.message || error.text : error;
+export function drawImpaled(error: CustomError | Error): void {
+  const errorCode =
+    "code" in error || "status" in error ? (error as CustomError).code || (error as CustomError).status : "xxx";
+  const errorString = error.message || String(error);
   console.log(error);
   let terminalColour = terminalColours();
   console.log(`---------------------------------------`);