+++ src/index.ts
@@ -1,5 +1,6 @@
 import { drawImpaled } from "./drawings/impaledDrawing";
+import { CustomError } from "./interfaces/ErrorTypes";
 
-export default function errorDrawing(error: any): void {
+export default function errorDrawing(error: CustomError | Error): void {
   drawImpaled(error);
 }