+++ src/test/test.ts
@@ -1,3 +1,7 @@
 import errorDrawing from "..";
 
-errorDrawing({ code: 500, message: "test" });
+//Custom Error interface
+errorDrawing({ code: 500, message: "Custom test arg error" });
+
+// A normal error
+errorDrawing(new Error("This is a 'new Error'-error"));